Decompile func00013a40

+++ src/include/lib/lib_13900.h
@@ -10,7 +10,7 @@ u32 func00013980(void);
 void func0001398c(s32 value);
 void func000139c8(void);
 s32 func000139e8(void);
-void func00013a40(u32 arg0);
+void func00013a40(u8 arg0);
 void func00013ab8(u32 arg0);
 void func00013c4c(s8 index);
 void func00013c94(void);
